Move Fiber's instrumentation check into fibers package.
